Implements cache memoization (#6)
Adds the ability to memoize cached values, reducing the number of times values are resolved.

Introduces a `hasMemoization` method in the `CacheEntity` class to determine whether to use memoization.

Refactors the `HasCacheMethods` trait to use the `memo` method of the cache driver when `hasMemoization` returns `true`.

Increases phpstan level to 9 for improved code quality.